+++ OmniUI/omnicore/src/OmniQmlRegistration.cpp
@@ -18,7 +18,6 @@
 #include "OmniTilePane.h"
 #include "OmniImGuiCanvas.h"
 #include "OmniPlot.h"
-#include "OmniAudioGraph.h"
 #include "OmniJuceView.h"
 #include "OmniDockArea.h"
 #include "OmniSplitView.h"
@@ -111,7 +110,6 @@ void OmniUI::registerQmlTypes() {
     qmlRegisterType<OmniTilePane>("OmniLayout", 1, 0, "TilePane");
 
     // Audio
-    qmlRegisterType<OmniAudioGraph>("OmniAudio", 1, 0, "AudioGraph");
     qmlRegisterType<OmniMidiHandler>("OmniAudio", 1, 0, "MidiHandler");
     qmlRegisterType<OmniSequencer>("OmniAudio", 1, 0, "Sequencer");
     qmlRegisterType<OmniAudioPlayer>("OmniAudio", 1, 0, "AudioPlayer");